Locally stored some javascript and css files that were linked to a bootstrap website.
Added core.js and core.css to hide and reveal divs.

Added some functionality to the nav-bar using javascript (not jQuery)

Added event handlers to the links in the nav-bar.

Added a folder for all html files pertaining to the home of the website.

Used django templating to split up components of home.html to make it more manageable.